refactor: DRY improvements and file splitting for usage code
- Extract duplicated JSONB breakdown SQL into breakdownUpdateSQL() helper
- Split admin analytics queries (ConnectorUsage, AgentUsage) into
  db/usage_admin.go for better file organization
- Remove dead GetTopUsersByUsage/TopUserUsage code and tests (endpoint
  was removed in security fix)
- Consolidate period resolution into resolvePeriodParam() to eliminate
  repeated default-period boilerplate across handlers
