+++ db/usage_admin_test.go
@@ -9,77 +9,6 @@ import (
 	"github.com/supersuit-tech/permission-slip-web/db/testhelper"
 )
 
-func TestGetTopUsersByUsage(t *testing.T) {
-	t.Parallel()
-	tx := testhelper.SetupTestDB(t)
-	ctx := context.Background()
-
-	uid1 := testhelper.GenerateUID(t)
-	uid2 := testhelper.GenerateUID(t)
-	uid3 := testhelper.GenerateUID(t)
-	testhelper.InsertUser(t, tx, uid1, "u1_"+uid1[:8])
-	testhelper.InsertUser(t, tx, uid2, "u2_"+uid2[:8])
-	testhelper.InsertUser(t, tx, uid3, "u3_"+uid3[:8])
-
-	periodStart, periodEnd := db.BillingPeriodBounds(time.Now())
-
-	// uid1: 5 requests, uid2: 10 requests, uid3: 1 request
-	for i := 0; i < 5; i++ {
-		if _, err := db.IncrementRequestCount(ctx, tx, uid1, periodStart, periodEnd); err != nil {
-			t.Fatalf("increment uid1: %v", err)
-		}
-	}
-	for i := 0; i < 10; i++ {
-		if _, err := db.IncrementRequestCount(ctx, tx, uid2, periodStart, periodEnd); err != nil {
-			t.Fatalf("increment uid2: %v", err)
-		}
-	}
-	if _, err := db.IncrementRequestCount(ctx, tx, uid3, periodStart, periodEnd); err != nil {
-		t.Fatalf("increment uid3: %v", err)
-	}
-
-	results, err := db.GetTopUsersByUsage(ctx, tx, periodStart, 2)
-	if err != nil {
-		t.Fatalf("GetTopUsersByUsage: %v", err)
-	}
-	if len(results) != 2 {
-		t.Fatalf("expected 2 results, got %d", len(results))
-	}
-	if results[0].UserID != uid2 {
-		t.Errorf("expected top user to be uid2, got %s", results[0].UserID)
-	}
-	if results[0].RequestCount != 10 {
-		t.Errorf("expected 10 requests, got %d", results[0].RequestCount)
-	}
-	if results[1].UserID != uid1 {
-		t.Errorf("expected second user to be uid1, got %s", results[1].UserID)
-	}
-}
-
-func TestGetTopUsersByUsage_ClampLimit(t *testing.T) {
-	t.Parallel()
-	tx := testhelper.SetupTestDB(t)
-	ctx := context.Background()
-
-	periodStart, _ := db.BillingPeriodBounds(time.Now())
-
-	// Zero limit should default to 10, not panic.
-	results, err := db.GetTopUsersByUsage(ctx, tx, periodStart, 0)
-	if err != nil {
-		t.Fatalf("GetTopUsersByUsage(limit=0): %v", err)
-	}
-	if results == nil {
-		// nil is fine for empty result
-	}
-
-	// Negative limit should also be clamped.
-	results, err = db.GetTopUsersByUsage(ctx, tx, periodStart, -1)
-	if err != nil {
-		t.Fatalf("GetTopUsersByUsage(limit=-1): %v", err)
-	}
-	_ = results
-}
-
 func TestGetUsageByConnectorForUser(t *testing.T) {
 	t.Parallel()
 	tx := testhelper.SetupTestDB(t)
